requirements.txt
torch==2.1.2
torchvision==0.16.2   
transformers==4.41.0
lightning==2.2.4
deepspeed==0.14.2
datasets==2.19.2
langcahin
langchain-huggingface
faiss-cpu
rank_bm25
langchainhub